+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 All notable changes to this project are documented in this file.
 
-## 2025-07-05
+## 2025-07-17
+### Added
+- Added editing functionality for workspaces in `Workspaces.js` component.
 ### Changed
 - Refactored workspace name handling in `Workspaces.js` to improve variable naming and initialization for workspace creation and editing functions.